+++ lib/application_controller.dart
@@ -1,4 +1,4 @@
-import 'package:chewie_with_danmaku/src/flutter_danmaku_controller.dart';
+import 'package:chewie_with_danmaku/flutter_danmaku/flutter_danmaku_controller.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'custom_chewie/chewie.dart';
@@ -21,5 +21,13 @@ class ApplicationController extends GetxController with WidgetsBindingObserver {
 
   void setController(FlutterDanmakuController controller) {
     flutterDanmakuController = controller;
+  }
+
+  @override
+  void dispose() {
+    videoPlayerController.dispose();
+    chewieController.dispose();
+    flutterDanmakuController.dispose();
+    super.dispose();
   }
 }